Physics Layer Freeze Bug
I think the reason for the bug was 2 things:
1) It's easier to set the _editMode as YES in the init of a new physics
layer because for now when we reset level we make a whole new physics
layer anyway.  When Claire makes the other type of reset button, she'll
of course have to set _editMode to yes there.
2) MORE IMPORTANT: Sometimes the ball body was being destroyed upon
"Get the Ball Rolling" being pressed because it was created slightly
out of bounds.  This locked edit mode (because the ball was supposed to
be rolling), but the body was of course destroyed and so nothing
happened.  This issue I think has been fixed (I just moved the ball
slightly to the right), but let me know if you still end up
encountering it!